move generate main into the post gen code directly
